refactor(api): transform 헬퍼 common 으로 lift + about mapper bySortOrder 추출
두 가지 cleanup 일괄:

1. trimIfString / trimToNull / trimArray 가 projects + about 두 DTO 에 복사
   구현되어 있어 apps/api/src/common/dto/transforms.ts 로 lift. 두 DTO 가
   import. 시그니처/동작 동일 — 기존 spec 그대로 통과.

2. about mapper 의 정렬 lambda 가 6 곳 인라인 — project-detail.mapper.ts 의
   bySortOrder<T> 추출 패턴을 그대로 미러링해 1 줄 정의 + 6 번 재사용.

api 27 suite / 128 test 그린 확인.

Closes #127
Refs #70
